requirements.txt
Flask==1.1.1
newsapi-python==0.2.6
google-cloud-storage
